fixing circular deps between asset-swapper/types and asset-swapper/utils/quote_report_generator
